devhub/requirements.txt
# Core dependencies
openai>=1.0.0
chromadb>=0.4.0

# OpenTelemetry (for Session 1 - students will use these)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0

# Testing (for Session 2 - students will use these)
deepeval>=3.0.0
pytest>=7.0.0

# LLM Observability (for Session 4 - students will use these)
langfuse>=3.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
